+++ apps/fluent-tester/src/TestComponents/Badge/BasicBadgeTest.tsx
@@ -1,6 +1,6 @@
 /* eslint-disable @typescript-eslint/no-var-requires */
-import React, { useState, useCallback } from 'react';
-import { View, Platform, Text, Image, FlexStyle } from 'react-native';
+import React, { useState, useCallback, useMemo } from 'react';
+import { View, Platform, Text, Image } from 'react-native';
 import {
   Badge,
   BadgeAppearance,
@@ -12,12 +12,15 @@ import {
   BadgeSize,
   BadgeSizes,
   BadgeIconPosition,
+  BadgeProps,
+  BadgeTokens,
 } from '@fluentui-react-native/badge';
 import { SvgIconProps } from '@fluentui-react-native/icon';
 import { StyledPicker } from '../Common/StyledPicker';
 import { satyaPhotoUrl } from './../PersonaCoin/styles';
 import TestSvg from '../../FluentTester/test-data/test.svg';
 import { ToggleButton } from '@fluentui/react-native';
+import { useFluentTheme } from '@fluentui-react-native/framework';
 
 const badgeColors: BadgeColor[] = [...BadgeColors];
 const badgeShapes: BadgeShape[] = [...BadgeShapes];
@@ -25,18 +28,9 @@ const badgeSizes: BadgeSize[] = [...BadgeSizes];
 const badgeAppearances: BadgeAppearance[] = [...BadgeAppearances];
 const badgeIconPositions = ['before', 'after'];
 
-const StyledBadge = Badge.customize({
-  fontWeight: 'bold',
-  fontSize: 12,
-  fontFamily: 'Georgia',
-  backgroundColor: '#f09',
-  borderColor: 'purple',
-  color: 'yellow',
-  borderWidth: 4,
-  borderStyle: 'dashed',
-  borderRadius: 2,
-  iconColor: 'cyan',
-});
+const useCustomizedBadge = (tokensAndprops: BadgeProps & BadgeTokens) => {
+  return useMemo(() => Badge.customize({ ...tokensAndprops }), [tokensAndprops]);
+};
 
 export const BasicBadge: React.FunctionComponent = () => {
   const [badgeAppearance, setBadgeAppearance] = useState<BadgeAppearance>('filled');
@@ -45,13 +39,17 @@ export const BasicBadge: React.FunctionComponent = () => {
   const [size, setSize] = useState<BadgeSize>('medium');
   const [showIcon, setShowIcon] = useState(false);
   const [iconPosition, setIconPosition] = useState<BadgeIconPosition>('before');
+  const [showShadow, setShowShadow] = useState(false);
 
   const onBadgeAppearanceChange = useCallback((value) => setBadgeAppearance(value), []);
   const onBadgeColorChange = useCallback((value) => setBadgeColor(value), []);
   const onShapeChange = useCallback((value) => setShape(value), []);
   const onSizeChange = useCallback((value) => setSize(value), []);
   const onIconPositionChange = useCallback((value) => setIconPosition(value), []);
+  const onShowShadowChange = useCallback(() => setShowShadow(!showShadow), [showShadow, setShowShadow]);
+  const onShowIconChange = useCallback(() => setShowIcon(!showIcon), [showIcon, setShowIcon]);
 
+  const theme = useFluentTheme();
   const svgProps: SvgIconProps = {
     src: TestSvg,
     viewBox: '0 0 500 500',
@@ -65,13 +63,30 @@ export const BasicBadge: React.FunctionComponent = () => {
 
   const svgIconsEnabled = ['ios', 'macos', 'win32', 'android'].includes(Platform.OS as string);
   const iconProps = { svgSource: svgProps, width: 20, height: 20 };
+
+  const shadow = showShadow ? theme.shadows.shadow4 : undefined;
+  const CustomBadge = useCustomizedBadge({ shadowToken: shadow });
+
   const badgeConfig = {
     appearance: badgeAppearance,
     badgeColor,
-    position: 'absolute' as FlexStyle['position'],
     size,
     shape,
   };
+
+  const StyledBadge = useCustomizedBadge({
+    fontWeight: 'bold',
+    fontSize: 12,
+    fontFamily: 'Georgia',
+    backgroundColor: '#f09',
+    borderColor: 'purple',
+    color: 'yellow',
+    borderWidth: 4,
+    borderStyle: 'dashed',
+    borderRadius: 2,
+    iconColor: 'cyan',
+    shadowToken: theme.shadows.shadow16,
+  });
 
   return (
     <View>
@@ -81,27 +96,29 @@ export const BasicBadge: React.FunctionComponent = () => {
       <StyledPicker prompt="Size" selected={size} onChange={onSizeChange} collection={badgeSizes} />
       {svgIconsEnabled && (
         <>
-          <ToggleButton onClick={() => setShowIcon(!showIcon)} checked={showIcon}>
+          <ToggleButton onClick={onShowIconChange} checked={showIcon}>
             Set {showIcon ? ' Hide icon' : ' Show icon'}
           </ToggleButton>
           <StyledPicker prompt="Icon position" selected={iconPosition} onChange={onIconPositionChange} collection={badgeIconPositions} />
         </>
       )}
+      <ToggleButton onClick={onShowShadowChange} checked={showShadow}>
+        Set {showShadow ? ' Hide shadow' : ' Show shadow'}
+      </ToggleButton>
 
       <View style={{ position: 'relative', backgroundColor: 'yellow', padding: 20, width: 200 }}>
         <Text>Parent component for the Badge</Text>
         {svgIconsEnabled && showIcon ? (
-          <Badge {...badgeConfig} icon={{ svgSource: svgProps }} iconPosition={iconPosition}>
+          <CustomBadge {...badgeConfig} icon={{ svgSource: svgProps }} iconPosition={iconPosition}>
             Basic badge
-          </Badge>
+          </CustomBadge>
         ) : (
-          <Badge {...badgeConfig}>Basic badge</Badge>
+          <CustomBadge {...badgeConfig}>Basic badge</CustomBadge>
         )}
       </View>
-
       <Text>Size</Text>
-      <Badge size="tiny" shape="circular" />
-      <Badge size="extraSmall" shape="circular" badgeColor="red" />
+      <Badge size="tiny" />
+      <Badge size="extraSmall" badgeColor="red" />
       <Badge size="small">Small</Badge>
       <Badge size="medium">Medium</Badge>
       <Badge size="large">Large</Badge>