+++ packages/deprecated/theming-react-native/src/BaselinePlatformDefaults.ts
@@ -1,6 +1,9 @@
 import { ITheme } from '@uifabricshared/theming-ramp';
 import { defaultFluentTheme } from '@fluentui-react-native/default-theme';
 
+/**
+ * @deprecated
+ */
 export function getBaselinePlatformTheme(): ITheme {
   return defaultFluentTheme;
 }