+++ packages/experimental/RadioGroup/src/RadioGroup/RadioGroup.tsx
@@ -15,7 +15,9 @@ export const RadioGroup = compose<RadioGroupType>({
   ...stylingSettings,
   slots: {
     root: View,
-    label: Text,
+    label: View,
+    labelText: Text,
+    required: Text,
     container: FocusZone,
   },
   useRender: (userProps: RadioGroupProps, useSlots: UseSlots<RadioGroupType>) => {
@@ -28,15 +30,22 @@ export const RadioGroup = compose<RadioGroupType>({
         return null;
       }
 
-      const { label, defaultTabbableElement, isCircularNavigation, ...mergedProps } = mergeProps(radioGroup.props, final);
+      const { label, required, defaultTabbableElement, isCircularNavigation, ...mergedProps } = mergeProps(radioGroup.props, final);
+
+      const labelComponent = (
+        <Slots.label>
+          <Slots.labelText key="label">{label}</Slots.labelText>
+          {!!required && <Slots.required>{'*'}</Slots.required>}
+        </Slots.label>
+      );
 
       // Populate the buttonKeys array
       if (children) {
         // eslint-disable-next-line @typescript-eslint/ban-ts-comment
         // @ts-ignore - TODO, fix typing error
-        contextValue.buttonKeys = React.Children.map(children, (child: React.ReactChild) => {
+        contextValue.values = React.Children.map(children, (child: React.ReactChild) => {
           if (React.isValidElement(child)) {
-            return child.props.buttonKey;
+            return child.props.value;
           }
         });
       }
@@ -44,7 +53,7 @@ export const RadioGroup = compose<RadioGroupType>({
       return (
         <RadioGroupProvider value={contextValue}>
           <Slots.root {...mergedProps}>
-            {label && <Slots.label>{label}</Slots.label>}
+            {label && labelComponent}
             <Slots.container isCircularNavigation defaultTabbableElement={defaultTabbableElement}>
               {children}
             </Slots.container>