+++ packages/experimental/RadioGroup/src/RadioGroup/RadioGroup.types.ts
@@ -2,6 +2,7 @@ import type { IViewProps } from '@fluentui-react-native/adapters';
 import { TextProps } from '@fluentui-react-native/text';
 import { IForegroundColorTokens, FontTokens } from '@fluentui-react-native/tokens';
 import { FocusZoneProps } from '@fluentui-react-native/focus-zone';
+import { ColorValue, ViewStyle } from 'react-native';
 
 export const radioGroupName = 'RadioGroup';
 
@@ -27,12 +28,27 @@ export interface RadioGroupState extends RadioGroupProps {
   selectedButtonRef?: React.MutableRefObject<any>;
 
   /**
-   * Array of radio button keys in the group
+   * Array of radio keys in the group
    */
-  buttonKeys?: string[];
+  values?: string[];
 }
 
-export interface RadioGroupTokens extends IForegroundColorTokens, FontTokens {}
+export interface RadioGroupTokens extends IForegroundColorTokens, FontTokens {
+  /**
+   * Color of required indicator
+   */
+  requiredColor?: ColorValue;
+
+  /**
+   * Amount of padding between the end of the label and the start of the required text
+   */
+  requiredPadding?: ViewStyle['padding'];
+
+  /**
+   * States that can be applied to a RadioGroup
+   */
+  disabled?: RadioGroupTokens;
+}
 
 export interface RadioGroupProps extends Pick<FocusZoneProps, 'isCircularNavigation' | 'defaultTabbableElement'>, IViewProps {
   /**
@@ -53,6 +69,16 @@ export interface RadioGroupProps extends Pick<FocusZoneProps, 'isCircularNavigat
   value?: string;
 
   /**
+   * Require a selection in this group. Adds the 'required' prop to all child Radio items.
+   */
+  required?: boolean;
+
+  /**
+   * Sets the RadioGroup to disabled if true
+   */
+  disabled?: boolean;
+
+  /**
    * Callback for receiving a notification when the choice has been changed
    */
   onChange?: (key: string) => void;
@@ -65,7 +91,9 @@ export interface RadioGroupInfo {
 
 export interface RadioGroupSlotProps {
   root: React.PropsWithRef<IViewProps>;
-  label?: TextProps;
+  label?: IViewProps;
+  labelText?: TextProps;
+  required?: TextProps;
   container: FocusZoneProps;
 }